Query BLM library with the given search parameters

Query BLM library with the given search parameters

blm_query(
  query = NULL,
  years = NULL,
  mediatypes = NULL,
  subjects = NULL,
  creators = NULL,
  languages = NULL
)

Arguments

query
string, no vector

years
string

mediatypes
string

subjects
string

creators
string

languages
string

Value

dataframe title, link

Examples

blm_query(query = "wildlife", subjects = c("Wildlife conservation", "land"))

#> [1] "https://archive.org/details/blmlibrary?&and[]=wildlife&sin=&and[]=subject:\"Wildlife+conservation\"&and[]=subject:\"land\""

#>                                                                               link
#> 1                                                  /details/skullsantlershor00vann
#> 2                                                  /details/wildlifeguildsin00shor
#> 3                                                  /details/easternstatesend13meri
#> 4                                                  /details/studiesforwildli12baum
#> 5                                                  /details/utahfishwildlife00unit
#> 6                                                  /details/endangeredwildli23meri
#> 7                                                  /details/southwestglendiv43gnia
#> 8                                                  /details/bureauoflandmana00abig
#> 9                                                 /details/rangewildlifetec005unse
#> 10                                                 /details/wildlifeeconomic00davi
#> 11                                                 /details/statusreportwild24unit
#> 12                                                 /details/watchablewildlif00unit
#> 13                                                 /details/livestockwildlif85sche
#> 14                                                 /details/wildlifefisherie15unit
#> 15                                                 /details/wildlifeeconomic07davi
#> 16                                               /details/1971annualwildli00unit_0
#> 17                                                 /details/fishwildlifemana00unse
#> 18                                                 /details/fishwildlife200000unit
#> 19                                                 /details/terrestrialwildl00mart
#> 20                                                 /details/arizonafishwildl00unit
#> 21                                                     /details/fishwildlife27unit
#> 22                                                 /details/1971annualwildli00unit
#> 23                                                 /details/nevadafishwildli18unit
#> 24                                                 /details/cooperativewildl00unse
#> 25                                                     /details/fishwildlife10unit
#> 26                                                 /details/fishwildlifeplan84unit
#> 27                                                 /details/eastmesawildlife00unit
#> 28                                               /details/careermanagement7800unit
#> 29                                               /details/fishwildlife200000unit_0
#> 30                                                 /details/guidelinesforwil32stan
#> 31                                                 /details/pariakanabcreekh09slon
#> 32                                                 /details/reductionofwildl00grov
#> 33                                                 /details/fishwildlifeplan83unit
#> 34                                                 /details/rileyridgenatura07envi
#> 35                                               /details/wildlifehabitats00unse_0
#> 36                                               /details/evaluationofbure4463unit
#> 37                                                 /details/wildlifehabitati09coop
#> 38                                                     /details/fishwildlife23unit
#> 39                                                     /details/fishwildlife16unit
#> 40                                                 /details/reportofcomplian00bioe
#> 41                                                 /details/applegateriverwa00unse
#> 42                                                 /details/strategyplanfort25mccl
#> 43                                                 /details/wildliferesource04albe
#> 44                                                 /details/compatibilityofo05acei
#> 45                                               /details/effectsoflivesto7166unit
#> 46                                                 /details/brotherseisareai00unse
#> 47                                               /details/wildlifeconserva00unit_0
#> 48                                                 /details/accomplishmentso00unit
#> 49                                                 /details/wildlifehabitato00unse
#> 50 /details/StudyEvalOilGasDevWildCharWildlifeResFederalLandsCentralArcticAreaofAK
#> 51                                                 /details/roancreekhabitat39unit
#> 52                                                 /details/conservationasse00conn
#> 53                                                 /details/nongamespeciesli40slsn
#> 54                                                 /details/progressreporton00nels
#> 55                                               /details/nutritionforagin00esqu_0
#> 56                                                 /details/nutritionforagin00esqu
#> 57                                               /details/draftenvironment00unit_2
#> 58                                                 /details/beaverdamintensi45unit
#> 59                                               /details/amargosacanyondu5986pupf
#> 60                                                 /details/salemdistrictpro00unit
#> 61                                                   /details/proposedpubliclaunit
#> 62                                                 /details/tuleelkincalifor00unit
#> 63                                                 /details/reporttobureauof00wagn
#> 64                                                 /details/proposedpublicla00unit
#> 65                                                 /details/recordofdecision13unit
#> 66                                                 /details/endangeredthreat52orda
#> 67                                                 /details/coastalplainoilg03unse
#> 68                                                 /details/surveyformountai00kein
#> 69                                               /details/coastalplainoilg02unse_1
#> 70                                               /details/recordofdecisionan00unit
#> 71                                               /details/coastalplainoilg02unse_0
#> 72                                               /details/coastalplainoilg01unse_1
#> 73                                               /details/coastalplainoilg01unse_0
#> 74                                                 /details/mountdomehabitat19lehm
#> 75                                               /details/dietdietselectio7126esqu
#>                                                                                                                                                                                                                           title
#> 1                                                                                                                                                             Skulls, antlers, and horns ; Wildlife trees ; Wildlife biologists
#> 2                                                                                                                                                                                    Wildlife guilds in Arizona desert habitats
#> 3                                                                                                                                                                                            Eastern States endangered wildlife
#> 4                                                                                                                                                                                          Studies for wildlife on energy areas
#> 5                                                                                                                                                   Utah fish and wildlife 2000 : a statewide strategy for the wildlife program
#> 6                                                                                                                                                                                          Endangered wildlife : Eastern States
#> 7                                                                                                                                                                                Southwest Glendive wildlife baseline inventory
#> 8                                                                                                                                                         Bureau of Land Management : FY 1996 fisheries and wildlife assessment
#> 9                                                                                                                                                                                         Range and wildlife technical workshop
#> 10                                                                                                                                                                           Wildlife economics and productivity analysis guide
#> 11                                                                                                                                                                               Status report : wildlife and fisheries program
#> 12                                                                                                                                                                   Watchable wildlife strategy for California Desert District
#> 13                                                                                                                                        Livestock-wildlife interactions on public lands in New Mexico : a bibliographic study
#> 14                                                                                                                                                                                    Wildlife and Fisheries Information System
#> 15                                                                                                                                                                 Wildlife economics and productivity analysis : student guide
#> 16                                                                                                                                                                      1971 annual [wildlife] report : Alaska [through] Nevada
#> 17                                                                                                                                                                                            Fish and wildlife management plan
#> 18                                                                                                                                                                             Fish and wildlife 2000 : a vision for the future
#> 19                                                                                                                                              Terrestrial wildlife inventory in selected coal areas of Montana : final report
#> 20                                                                                                                                                                        Arizona fish and wildlife 2000 : designing our future
#> 21                                                                                                                                                                               Fish and wildlife 2000 : a plan for the future
#> 22                                                                                                                                                                 1971 annual [wildlife] report : New Mexico [through] Wyoming
#> 23                                                                                                                                                                        Nevada Fish and Wildlife 2000 : looking to the future
#> 24                                                                                                                                           Cooperative Wildlife Program : phase I literature search and review project report
#> 25                                                                                                                                                              Fish & wildlife 2000 : annual report of accomplishments FY 1991
#> 26                                                                                                                                   Fish and wildlife plan for Colorado : biennial report of accomplishments, FY 1991 and 1992
#> 27                                                                                                                                               East Mesa wildlife habitat management plan : CA-06WHA-70 : a Sikes Act project
#> 28                                                                                                                                                          Career Management Team findings for wildlife & fisheries biologists
#> 29                                                                                                                                                    Fish & wildlife 2000 : annual report of accomplishments, fiscal year 1994
#> 30                                                                                                                                                     Guidelines for wildlife habitat improvements in Oregon grazing districts
#> 31                                                                                                                                                         Paria-Kanab Creek habitat management plan : wildlife habitat area-T6
#> 32                                                                                                                                  The reduction of wildlife mortality in the sump pits of southeast New Mexico : final report
#> 33                                                                                                                                                               A Fish and wildlife plan for Colorado : program for the decade
#> 34                                                                                                                                                    Riley Ridge natural gas project : wildlife and fisheries technical report
#> 35                                                                                                                                             Wildlife habitats in managed rangelands : the Great Basin of southeastern Oregon
#> 36                                        Evaluation of Bureau of Land Management wildlife program interactions with rangeland management : a report to the Bureau of Land Management, United States Department of the Interior
#> 37                                                                                                                                           Wildlife habitat inventory, Kaiparowits powerline right of way, Riverside District
#> 38                                                                                                                                                                               Fish and wildlife 2000 : a plan for the future
#> 39                                                                                                                                                               Fish & wildlife 2000 : annual progress report fiscal year 1990
#> 40 Report of compliance for mitigation requirements for sensitive wildlife and special status plant species during seismic operations on the Nuevo Energy Company Cymric 2D seismic survey, McKittrick, Kern County, California
#> 41                                                                                                                                          Applegate River Watershed assessment : aquatic, wildlife, and special plant habitat
#> 42                                                                                                                         Strategy plan for training personnel in BLM's wildlife, fisheries, and special status plants program
#> 43                                                                                                                                     Wildlife resource inventory of the Chaco strippable coal area, New Mexico : final report
#> 44                                                                                                                            Compatibility of oil and gas development in the north Salton Sea with fish and wildlife resources
#> 45                                                                                                                          Effects of livestock grazing on wildlife, watershed, recreation and other resource values in Nevada
#> 46                                                                                                                                    Brothers EIS area in Central Oregon : wildlife habitat interrelationships : working draft
#> 47                                                                                                                Wildlife conservation & our hunting heritage : distribution of elk, mule deer, bighorn sheep, and wild turkey
#> 48                                                                                            Accomplishments of the fisheries, wildlife, threatened and endangered species, and plant conservation programs : fiscal year 2011
#> 49                                                                         Wildlife habitat & our shared conservation stewardship : distribution of elk, mule deer, bighorn sheep, and wild turkey in the western United States
#> 50                                                                          Study to evaluate oil and gas development, wilderness characteristics, and wildlife resources on Federal lands in the central Arctic area of Alaska
#> 51                                                                                                                                                                                           Roan Creek habitat management plan
#> 52                                                                                                                                                        Conservation assessment of greater sage-grouse and sagebrush habitats
#> 53                                                                                           Nongame species literature search in support of wildlife inventories in the Elko, Ely, and Battle Mountain BLM Districts of Nevada
#> 54                                                                                        Progress report on the BLM in California : text of talk by Neal D. Nelson : California Wildlife Federation, San Rafael, June 30, 1962
#> 55                                                                                                                                                Nutrition and foraging ecology of the desert tortoise : FY 1990 annual report
#> 56                                                                                                                                                Nutrition and foraging ecology of the desert tortoise : FY 1989 annual report
#> 57                                                                                                                Draft environmental impact statement : Florida Canyon Mine, mine expansion and comprehensive reclamation plan
#> 58                                                                                                                                     Beaver Dam intensive inventory, habitat management plan and final environmental analysis
#> 59                                                                                                                                                                           Amargosa Canyon-Dumont Dunes proposed natural area
#> 60                                                                                                                                                                                     Salem District project update: June 2005
#> 61                                                                                   Proposed public land withdrawal : Nellis Air Force bombing range, Nye, Clark, and Lincoln counties, Nevada : draft environmental statement
#> 62                                                                                                                                                                         The tule elk in California : 10th report to Congress
#> 63                                                                            Report to the Bureau of Land Management and U.S. Forest Service on raptors inhabiting Brown's Park and the Flaming Gorge National Recreation Area
#> 64                                                                            Proposed public land withdrawal : Nellis Air Force bombing range, Nye, Clark, and Lincoln counties, Nevada : final environmental impact statement
#> 65                                                                                                                                                      Record of decision : resource management plan for Redding Resource Area
#> 66                                                                                                                                           Endangered and threatened fauna survey of Dona Ana and Sierra counties, New Mexico
#> 67                                                                                                                                             Coastal plain oil and gas leasing program, environmental impact statement, final
#> 68                                                                                                                                  Survey for Mountain Plover (Charadrius montanus) on Federal Lands in the Powder River Basin
#> 69                                                                                                                                             Coastal plain oil and gas leasing program, environmental impact statement, final
#> 70                                                                                                                                    Record of decision and resource management plan for the Winnemucca District Planning Area
#> 71                                                                                                                                             Coastal plain oil and gas leasing program, environmental impact statement, final
#> 72                                                                                                                                             Coastal plain oil and gas leasing program, environmental impact statement, final
#> 73                                                                                                                                             Coastal plain oil and gas leasing program, environmental impact statement, final
#> 74                                                               Mount Dome habitat management plan : a plan for the protection and management of a bald eagle wintering and nesting area on public land in northern California
#> 75                                                                                                                           Diet and diet selection of the desert tortoise (Gopherus agassizii) in the northeast Mojave Desert