Perform a search on Treesearch's online interface

Perform a search on Treesearch's online interface

treesearch_web_query(
  keywords = "",
  authorlastname = "",
  yearfrom = "",
  yearto = "",
  station = "",
  series = "",
  volume = "",
  treesearch_url = "https://www.fs.usda.gov/treesearch/search?"
)

Arguments

keywords
string

authorlastname
string

yearfrom
string

yearto
string

station
string

series
string

volume
string

treesearch_url
string optional argument

Value

dataframe title,id,text for top 20 results (or all if fields blank).

Examples

treesearch_web_query(keywords = "increment-borer methods fire history", 
                 authorlastname <- "Arno")

#> getting page 1 

#>                                                                                                                                                                         title
#> 1                                                     Illustrated guide to the emerald ash borer, Agrilus planipennis Fairmaire and related species (Coleoptera, Buprestidae)
#> 2                                                                                                                     Reducing borer damage in oak regeneration and sawtimber
#> 3                                                                                         The history and future of methyl bromide alternatives in the southern United States
#> 4                                              The history and future of methyl bromide alternatives used in the production of forest seedlings in the southern United States
#> 5                          Climate change may increase the drought stress of mesophytic trees downslope with ongoing forest mesophication under a history of fire suppression
#> 6               Environmental influences on tree and stand increment, proceedings of the mensuration, growth and yield instruments and methods in forest mensuration workshop
#> 7                                                                             Evaluating fuel complexes for fire hazard mitigation planning in the southeastern United States
#> 8                                                                                                                              Emerald ash borer biology and invasion history
#> 9                                                                                                                             History of emerald ash borer biological control
#> 10                                                                 Mapping upland hardwood site quality and productivity with GIS and FIA in the Blue Ridge of North Carolina
#> 11                                                                                      Validation of annual growth rings in freshwater mussel shells using cross dating .Can
#> 12                                                   Long-term effects of fire and fire-return interval on population structure and growth of longleaf pine (Pinus palustris)
#> 13                                                                          The history of emerald ash borer discoveries in the Upper Peninsula of Michigan from 2005 to 2007
#> 14 Emergence of Buprestidae, Cerambycidae, and Scolytinae (Coleoptera) from mountain pine beetle-killed and fire-killed ponderosa pines in the Black Hills, South Dakota, USA
#> 15                                                                                           A dendrochronological fire history of Opeongo lookout in Algonquin Park, Ontario
#> 16                                                                                          Removing the tree-ring width biological trend using expected basal area increment
#> 17                      Variables associated with the occurrence of Ips beetles, red turpentine beetle and wood borers in live and dead ponderosa pines with post-fire injury
#> 18                                                                        Potential relative increment (PRI): a new method to empirically derive optimal tree diameter growth
#> 19                                                                                                                                 Oak Decline in Missouri: History Revisited
#> 20                                                                 Fire history and fire management implications in the Yukon Flats National Wildlife Refuge, interior Alaska
#>    treesearchid
#> 1         49163
#> 2         18627
#> 3         42732
#> 4         45620
#> 5         59964
#> 6         49833
#> 7         40556
#> 8         49254
#> 9         49321
#> 10        33365
#> 11        33530
#> 12        35455
#> 13        19372
#> 14        48598
#> 15        12339
#> 16        32501
#> 17        52313
#> 18         2407
#> 19         6557
#> 20        31543
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     text
#> 1                                                                                                                                                                                                                                                               Illustrated guide to the emerald ash borer, Agrilus planipennis Fairmaire and related species (Coleoptera, Buprestidae)The 33 species of Agrilus (Coleoptera: Buprestidae) hypothesized to be most closely related or most similar to Agrilus planipennis Fairmaire (the emerald ash borer), are described and illustrated. Morphology (adults and immatures), biology, distribution, detailed taxonomic history and systematics are presented for each species,...Author(s):M. Lourdes Chamorro; Eduard Jendek; Robert A. Haack; Toby Petrice; Norman E. Woodley; Alexander S. Konstantinov; Mark G. Volkovitsh; Xing-Ke Yang; Vasily V. GrebennikovYear:2015Keywords:Coleoptera, Buprestidae, Agrilus planipennis, emerald ash borerSource:Sofia, Bulgaria: Pensoft Publishers. 197 p.
#> 2                                                                                                                                                                                                                                                                                                                                                                                             Reducing borer damage in oak regeneration and sawtimberBorers cause millions of dollars in damaged wood annually to oak stands, and adversely affect the form and vigor of oak regeneration. A moth and four species of beetles cause most of the damage; the carpenterworm moth, the oak timberworm, the red oak borer, the living-beech borer, and the white oak borer. The larvae of these insects chew holes in the wood ranging from...Author(s):Jimmy R. GalfordYear:1989Keywords:Source:In: Hutchinson, Jay G., ed. Central hardwood notes. St. Paul, MN.: U.S. Dept. of Agriculture, Forest Service, North Central Forest Experiment Station. 8.09
#> 3                                                                                                                                                                                           The history and future of methyl bromide alternatives in the southern United StatesThis article gives a brief history of the efforts of the Southern Forest Nursery Management Cooperative (SFNMC) in testing methyl bromide (MBr) alternatives for soil fumigation. In the southeastern United States, fumigation with MBr has been the most commonly used method for producing high quality, pest-free forest seedlings in an environment that is conducive for soil...Author(s):Tom E. StarkeyYear:2012Keywords:soil fumigation, chloropicrin, loblolly pine, high barrier plastics, broadcast fumigationSource:In: Haase, D. L.; Pinto, J. R.; Riley, L. E., tech. coords. National Proceedings: Forest and Conservation Nursery Associations - 2011. Proc. RMRS-P-68. Fort Collins, CO: USDA Forest Service, Rocky Mountain Research Station. p. 31-35.
#> 4                                                                                                  The history and future of methyl bromide alternatives used in the production of forest seedlings in the southern United StatesThis paper gives a brief history of the Southern Forest Nursery Management Cooperative’s (SFNMC) efforts in testing methyl bromide (MBr) alternatives for soil fumigation. In the southeastern United States, fumigation with MBr has been the most commonly used method for producing high quality, pest-free forest-tree seedlings in an environment that is conducive for soil-...Author(s):Scott A. EnebakYear:2013Keywords:soil fumigation, chloropicrin, loblolly pine, high barrier plastics, broadcast fumigatioSource:In: Haase, D. L.; Pinto, J. R.; Wilkinson, K. M., technical coordinators. National Proceedings: Forest and Conservation Nursery Associations - 2012. Proceedings RMRS-P-69. Fort Collins, CO: U.S. Department of Agriculture, Forest Service, Rocky Mountain Research Station. p. 20-25.
#> 5                                                                                                                                                                                                                                 Climate change may increase the drought stress of mesophytic trees downslope with ongoing forest mesophication under a history of fire suppressionIn mountainous headwater catchments, downslope flow of subsurface water could buffer downslope forest communities from soil moisture stress during drought. Here we investigated changes in landscape-scale vegetation patterns at five forested headwater catchments in the Coweeta Hydrologic Laboratory in the southern Appalachians. We used a ca. 30-year Landsat...Author(s):Taehee Hwang; Lawrence E. Band; Chelcy F. Miniat; James M. Vose; Jennifer D. Knoepp; Conghe Song; Paul V BolstadYear:2020Keywords:Forest mesophication, fire suppression, isohydricity and anisohydricity, forest hydrology, droughtSource:Frontiers in Forests and Global Change
#> 6                                                                                                                                                                                                                                                                                                    Environmental influences on tree and stand increment, proceedings of the mensuration, growth and yield instruments and methods in forest mensuration workshopAn international conference on Environmental Influences on Tree and Stand Increment was held at the University of New Hampshire, Durham, New Hampshire, USA between September 23 and 27, 1985. The meeting included recommendations from a previous I.U.F.R.O. meeting in Vienna, Austria. Demands for forest resources are increasing but the forest land base is constantly...Author(s):Dale S. Solomon; Thomas B. BrannYear:1986Keywords:Source:1985 September 23-27; Durham NH. Miscellaneous Publication 691. Maine Agricultural Experiment Station, University of Maine. 186 p. 
#> 7                                                                                                                                                                                                                                                                                      Evaluating fuel complexes for fire hazard mitigation planning in the southeastern United StatesFire hazard mitigation planning requires an accurate accounting of fuel complexes to predict potential fire behavior and effects of treatment alternatives. In the southeastern United States, rapid vegetation growth coupled with complex land use history and forest management options requires a dynamic approach to fuel characterization. In this study we assessed...Author(s):Anne G. Andreu; Dan Shea; Bernard R. Parresol; Roger D. OttmarYear:2012Keywords:fuel characteristics, surface fire behavior, fuel reduction, fuel load, Fuel Characteristic Classification System, southeastern United StatesSource:Forest Ecology and Management. 27: 4–16
#> 8                                                                                                                                                                                                                             Emerald ash borer biology and invasion historyThe emerald ash borer (EAB), Agrilus planipennis Fairmaire (Coleoptera: Buprestidae), is native to eastern Asia and is primarily a pest of ash (Fraxinus) trees (Fig. 1). Established populations of EAB were first detected in the United States and Canada in 2002 (Haack et al., 2002), and based on a dendrochronology study by Siegert...Author(s):Robert A. Haack; Yuri Baranchikov; Leah S. Bauer; Therese M. PolandYear:2015Keywords:emerald ash borer, Agrilus planipennis, pest, wood borer, ash, Fraxinus, invasive, exoticSource:In: Van Driesche, R.G.; Reardon, R.C., eds. Biology and control of emerald ash borer. FHTET-2014-09. Morgantown, WV: U.S. Department of Agriculture, Forest Service, Forest Health Technology Enterprise Team: 1-13. Chapter 1.
#> 9                                                                                                                                                                                                                                                                                                        History of emerald ash borer biological controlThe search for natural enemies of the emerald ash borer (EAB), Agrilus planipennis Fairmaire (Coleoptera: Buprestidae), in northeastern Asia, its native range, was initiated within a year of its discovery in the United States (Bauer et al., 2005, 2014). Although the official response to EAB’s invasion in both the United States and Canada...Author(s):Juli R. Gould; Leah S. Bauer; Jian J. Duan; David Williams; Houping LiuYear:2015Keywords:Source:In: Van Driesche, R.G.; Reardon, R.C., eds. Biology and control of emerald ash borer. FHTET-2014-09. Morgantown, WV: U.S. Department of Agriculture, Forest Service, Forest Health Technology Enterprise Team: 83-95. Chapter 5.
#> 10                                                                                                    Mapping upland hardwood site quality and productivity with GIS and FIA in the Blue Ridge of North CarolinaThe forested ecosystems of the southern Appalachians are some of the most diverse in North America due to the variability in climate, soils, and geologic parent material coupled with the complex topography found throughout the region. These same characteristics cause stands of upland hardwoods to be extremely variable with regard to site quality and productivity. Site...Author(s):Claudia A. Cotton; Stephen R. Prisley; Thomas R. FoxYear:2009Keywords:upland hardwood forests, site quality, site productivity, GIS, Blue RidgeSource:In: McWilliams, Will; Moisen, Gretchen; Czaplewski, Ray, comps. Forest Inventory and Analysis (FIA) Symposium 2008; October 21-23, 2008; Park City, UT. Proc. RMRS-P-56CD. Fort Collins, CO: U.S. Department of Agriculture, Forest Service, Rocky Mountain Research Station. 13 p.
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                     Validation of annual growth rings in freshwater mussel shells using cross dating .CanWe examined the usefulness of dendrochronological cross-dating methods for studying long-term, interannual growth patterns in freshwater mussels, including validation of annual shell ring formation. Using 13 species from three rivers, we measured increment widths between putative annual rings on shell thin sections and then removed age-related variation by...Author(s):Andrew L. Rypel; Wendell R. Haag; Robert H. FindlayYear:2009Keywords:Source:Can. J. Fish. Aquat. Sci., Vol. 65: 2224-2232
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                  Long-term effects of fire and fire-return interval on population structure and growth of longleaf pine (Pinus palustris)We investigated the effect of fire and fire frequency on stand structure and longleaf pine (Pinus palustris P. Mill.) growth and population demography in an experimental research area in a southwest Florida sandhill community. Data were collected from replicated plots that had prescribed fire-return intervals of 1, 2, 5, or 7 years or were left...Author(s):Chelcy R. Ford; Emily S. Minor; Gordon A. FoxYear:2010Keywords:Source:Can. J. For. Res. 40: 1410–1420
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The history of emerald ash borer discoveries in the Upper Peninsula of Michigan from 2005 to 2007Author(s):Andrew J. Storer; Jordan M. Marshall; Michael M. Philip; Deborah G. Mccullough; Nathan W. SiegertYear:2009Keywords:Source:In: McManus, Katherine A; Gottschalk, Kurt W., eds. Proceedings. 19th U.S. Department of Agriculture interagency research forum on invasive species 2008; 2008 January 8-11; Annapolis, MD. Gen. Tech. Rep. NRS-P-36. Newtown Square, PA: U.S. Department of Agriculture, Forest Service, Northern Research Station: 77.
#> 14                                                                                                                                                                                                                                                     Emergence of Buprestidae, Cerambycidae, and Scolytinae (Coleoptera) from mountain pine beetle-killed and fire-killed ponderosa pines in the Black Hills, South Dakota, USAWood borers (Coleoptera: Cerambycidae and Buprestidae) and bark beetles (Coleoptera: Curculionidae) infest ponderosa pines, Pinus ponderosa P. Lawson and C. Lawson, killed by mountain pine beetle (MPB), Dendroctonus ponderosae Hopkins, and fire. No data is available comparing wood borer and bark beetle densities or species guilds associated with MPB-killed or fire-...Author(s):Sheryl L. Costello; William R. Jacobi; Jose F. NegronYear:2013Keywords:wood borers, bark beetles, metallic wood-boring beetles, longhorn beetles, Dendroctonus ponderosaeSource:The Coleopterists Bulletin. 67(2): 149-154.
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                          A dendrochronological fire history of Opeongo lookout in Algonquin Park, OntarioFire-scarred stumps and natural snags of red pine (Pinus resinosa Ait.) on a 70-m tall bluff along Costello Creek, 2 km south of Lake Opeongo, were dated using dendrochronological methods. A fire scar chronology was constructed from 18 of the red pine samples that contained 34 dated fire scars.Author(s):Richard P. Guyette; Daniel C. DeyYear:1995Keywords:Fire history, dendrochronology, red pine, presettlement, disturbance, anthropogenicSource:Forest Research Report No. 134. 1995. pp. 1-4.
#> 16 Removing the tree-ring width biological trend using expected basal area incrementOne of the main elements of dendrochronological standardization is the removal of the biological trend, i.e., the progressive decline of ring width along a cross-sectional radius that is mostly caused by the corresponding increase in stem diameter over time. A very common option for removing this biological trend is to fit a modified negative exponential curve to the...Author(s):Franco Biondi; Fares QeadanYear:2008Keywords:long-term research, ponderosa pine, range research, silviculture, cultural resources, Fort Valley Experimental Forest, Long Valley Experimental Forest, http://www.rmrs.nau.edu/fortvalley/Source:In: Olberding, Susan D., and Moore, Margaret M., tech. coords. Fort Valley Experimental Forest-A Century of Research 1908-2008. Conference Proceedings; August 7-9, 2008; Flagstaff, AZ. Proc. RMRS-P-55. Fort Collins, CO: U.S. Department of Agriculture, Forest Service, Rocky Mountain Research Station. p. 124-131
#> 17                                                                                                                                                                                                               Variables associated with the occurrence of Ips beetles, red turpentine beetle and wood borers in live and dead ponderosa pines with post-fire injuryRecently, wildfires and prescribed burning have become more frequent in conifer forests of western North America. Most studies examining the impacts of insects on trees with post-fire injury have focused on contributions to tree mortality. Few studies have examined fire-caused injuries to estimate the probability of attack by insects. Scant data quantifying...Author(s):Jose F. Negron; Joel McMillin; Carolyn H. Sieg; James F. Fowler; Kurt K. Allen; Linda L. Wadleigh; John A. Anhold; Ken E. GibsonYear:2016Keywords:Dendroctonus valens, fire injury, Ips beetles, Ips spp., Pinus ponderosa, wood borersSource:Agricultural and Forest Entomology. 18: 313-326.
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Potential relative increment (PRI): a new method to empirically derive optimal tree diameter growthPotential relative increment (PRI) is a new method to derive optimal diameter growth equations using inventory information from a large public database. Optimal growth equations for 24 species were developed using plot and tree records from several states (Michigan, Minnesota, and Wisconsin) of the North Central US. Most species were represented by thousands of...Author(s):Don C BraggYear:2001Keywords:Source:Ecological Modelling. 137: 77-92.
#> 19                                                                                                                                                                                                                                                                                                                                                                                              Oak Decline in Missouri: History RevisitedIn the 1980s, following extreme winters in the late 1970s and severe droughts in 1976, 1980, and 1983, dead and dying scarlet and black oaks were found on 185,000 acres of the Mark Twain National Forest. That decline event was linked to environmental stresses (Law and Gott 1987). Severe oak decline is now affecting an estimated 500,000 acres on the Mark Twain. High-...Author(s):Jay R. Law; Ross Melick; Charly Studyvin; James R. SteinmanYear:2004Keywords:Source:Gen. Tech. Rep. SRS-73. Asheville, NC: U.S. Department of Agriculture, Forest Service, Southern Research Station. pp. 257-258
#> 20                                                                                                                                                                                                                                                                                                                                              Fire history and fire management implications in the Yukon Flats National Wildlife Refuge, interior AlaskaWe conducted this investigation in response to criticisms that the current Alaska Interagency Fire Management Plans are allowing too much of the landscape in interior Alaska to burn annually. To address this issue, we analyzed fire history patterns within the Yukon Flats National Wildlife Refuge, interior Alaska. We dated 40 fires on 27 landscape points within the...Author(s):S. A. Drury; P. J. GrissomYear:2008Keywords:fire history, fire management, interior Alaska, black spruce, white spruce, aspen, dendrochronologySource:Forest Ecology and Management. 256: 304­-312.